third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-overflow/scrollable-overflow-zero-one-axis-expected.txt

This is a testharness.js-based test.
PASS .scroll 1
FAIL .scroll 2 assert_equals: 
<div class="scroll zero-w" data-expected-scroll-width="100">
  <div class="zero-h"></div>
</div>
scrollWidth expected 100 but got 0
PASS .scroll 3
FAIL .scroll 4 assert_equals: 
<div class="scroll zero-w flex-col" data-expected-scroll-width="100">
  <div class="zero-h"></div>
</div>
scrollWidth expected 100 but got 1
FAIL .scroll 5 assert_equals: 
<div class="scroll zero-h grid" data-expected-scroll-height="100">
  <div class="zero-w"></div>
</div>
scrollHeight expected 100 but got 0
FAIL .scroll 6 assert_equals: 
<div class="scroll zero-w grid" data-expected-scroll-width="100">
  <div class="zero-h"></div>
</div>
scrollWidth expected 100 but got 1
Harness: the test ran to completion.
